Refactor: Extract SecurityHeadersMiddleware to dedicated file\n\nMoves the SecurityHeadersMiddleware class out of app/main.py into\napp/middleware/security.py to improve readability and separation of concerns.
